+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ### Added
 
+- Support local tiff files.
+- Support local zarr directories.
+
 ### Changed
 
 - Removed greedy matching of dimension order in Bio-Formats Zarr output. Just check if OME-Zarr.
 - Use WASM LZW decoder from `manzt/geotiff.js`
+- Bundle UMD build with deps.
 
 ## 0.4.2